extension: separete config opts to an interface
The workspace configuration options were being taken and checked in
activate() function code, mixing the data and logic in a quite messy way.
This commit separates the configuration options data in a separate interface
and populates it with default values merging what's retrieved from the
config file and an instance of the new interface with default values.

In this way, the logic and calling to the server are handled in the
activate() function without mixing information. The overall behavior didn't
change, but the organization deeply enhances the code reading and
understanding experience.