extension: add missing bits for 2.4.0 and warn DAP
Some configuration options were not being handled as needed for 2.4.0
version, like the ability of using Perl::LanguageServer from a container of
passing envvars to the client. This commit adds these missing bits.

Also, this commit add notes about the lack of DAP support on coc-nvim,
ensureing users are aware of it.