package: rename 'pls' to 'p::ls' to avoid confusion
There is another language server for Perl called "Perl Language Server" [1]
which uses the acronym PLS to refer to itself. There are not plans to
support it on coc-perl for now and the project seems quite silent for some
time already, but to avoid future problems this commit changes the
configuration options for Perl::LanguageServer [2] server to "p::ls" from
"pls" and also the source file name from "pls.ts" to "p_ls.ts". However, the
code itself didn't change from PLS, for the sake of simplicity. In case
something changes in the future wrt PLS support on coc-perl, the code must
be changed as well.

Now, instead of calling "perl.pls.enable: true" in coc-settings.json, you'll
need to pass "perl.p::ls.enable: true".