installer: add installation procedure
In order to make coc-perl a self-contained extension, we added a
installation step for installing Perl::LanguageServer to the user host
system in case the user chooses it as the server. The installation happens
through `cpan` and creates an item in the status bar for letting the user
know the server is being installed.

PerlNavigator installation support is coming soon.