README: be more clear about the automatic install process
coc-perl is able to install both supported language servers, but for doing
so the user need to first install coc-perl, then set which server he's
willing to use and, finally, restart both (neo)vim and open a Perl file or
issue `:CocRestart` in case he already sits on top of an open Perl file.
This process was quite unclear in the README file, thus this commit adds a
specific section about this behavior to make sure the user is aware of what
he needs to do.