navigator: fix server configuration mishandling
The PerlNavigator server only supports the `workspace/configuration` pull
request to get server's custom configuration done in the client side. So
far, coc-perl was using an old interface, that works with
Perl::LanguageServer, with PerlNavigator. As consiquence, the server was
using its own default values for each option.

This commit changes the behavior of coc-perl by replying to the pull request
from the server with a proper `DidChangedConfigurationNotification` method.