config: drop coc-perl own namespaces
When coc-perl began to support both Perl::LanguageServer and PerlNavigator
the namespace for Perl::LanguageServer (perl.*) seemed awkward due to its
overlap with possible "general" configurations for the coc-perl extension
itself, so it was decided to do the following move: create three different
namespaces:

1. perl.* : general coc-perl configs (we had none so far)
2. perl.p::ls.* : Perl::LanguageServer config options
3. perl.navigator.* : PerlNavigator config options

Even though the idea was interesting at first glence, it became a nightmare to
maintain once some bugs were found with PerlNavigator and the way it requests
the config options in a pull request, instead of the client sending it as a
push (old way still used in Perl::LanguageServer) operation. With that, the
default config namespace would've to be used (perlnavigator.*) instead of the
coc-perl specific one (perl.navigator.*). Commit 989195ce9da1 ("config:
aggregate PerlNavigator configs in the original namespace") tried to fix it.

However, the way the options are handled in the different spaces a user can
have its configurations (global (user home dir), workspace and default),
with automated settings file update, created a certain chicken and egg
situation where the fix would force a really unreliable and tricky code.

With that, I preferred to completely remove the "coc-perl specific namespaces"
support and only use the original ones.